GPII-1987: File renaming and tests fixing.
