+++ scripts/shared/authDataLoader.js
