+++ tests/scripts/data/prefs/franklin.md
