+++ tests/scripts/data/prefs/sammy.json
